Enforce YAML key ordering in yamllint
The yamllint key-ordering rule enforces ordering of mapping keys by
their Unicode point number, which is the same mechanism we're currently
enforcing ordering by.

Note that I'm specifically excluding the `.github` and `test`
directories from this rule, where this sort of enforcement doesn't make
as much sense.

At present, the existing key ordering check will continue to function,
but setting this in the yamllint config will allow ordering problems to
be surfaced by rosdistro-reviewer as well as any external editors that
are configured to look for a yamllint config.